+++ app/ids-modal/visible.js
@@ -1,4 +1,4 @@
-import IdsModal from '../../src/ids-modal/ids-modal';
+import IdsModal from '../../src/ids-modal';
 import IdsButton from '../../src/ids-button/ids-button';
 
 document.addEventListener('DOMContentLoaded', () => {