+++ src/ids-base/ids-xss-utils.d.ts
@@ -1,3 +1,9 @@
+/** Takes a string containing HTML and Text and removes any reference to console methods */
+export function sanitizeConsoleMethods(html: string): string;
+
+/** Takes a string containing HTML and Text and removes any functions or script tags */
+export function sanitizeHTML(html: string): string;
+
 /** Takes a string containing HTML and Text and removes all HTML tags */
 export function stripTags(html: string): string;