+++ src/ids-modal/index.js
@@ -1,3 +1,2 @@
 export { default } from './ids-modal';
 export { default as IdsOverlay } from './ids-overlay';
-export { default as IdsModalButton } from './ids-modal-button';